requirements.txt
pyzmq==2.2.0
meinheld
nose
